backend/requirements.txt
streamlit>=1.18.0
pandas>=1.5.0
numpy>=1.22.0
requests>=2.28.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
reportlab>=3.6.0
flask>=3.0.0
flask-cors>=5.0.0
flask-jwt-extended>=4.7.0
gunicorn>=23.0.0
flask-swagger-ui>=4.11.0
marshmallow>=3.26.0
apispec>=6.0.0
apispec-webframeworks>=0.5.2
marshmallow-sqlalchemy>=0.30.0
